move secrets methods to a separate file for clarity
